+++ CHANGELOG.md
@@ -13,6 +13,9 @@
   has been removed
 - Deprecated commands `pcs cluster remote-node add | remove` have been removed
   as they were replaced with `pcs cluster node add-guest | remove-guest`
+- `pcs resource create` no longer allows creating master resources as they are
+  deprecated in Pacemaker 2.x; use `pcs resource create ... promotable` or
+  `pcs resource create ... clone promotable=true` instead ([rhbz#1542288])
 
 ### Added
 - Validation for an unaccessible resource inside a bundle ([rhbz#1462248])
@@ -104,6 +107,8 @@
 - Pcs now configures corosync to put timestamps in its log ([rhbz#1615420])
 - Option `-V` has been replaced with `--full` and a CIB file can be specified
   only using option `-f` in `pcs cluster verify`
+- Master resources have been replaced with promotable clone resources in
+  accordance with changes in pacemaker-2.0 ([rhbz#1542288])
 
 ### Security
 - CVE-2018-1086: Debug parameter removal bypass, allowing information disclosure
@@ -120,6 +125,7 @@
 [rhbz#1488044]: https://bugzilla.redhat.com/show_bug.cgi?id=1488044
 [rhbz#1533866]: https://bugzilla.redhat.com/show_bug.cgi?id=1533866
 [rhbz#1536121]: https://bugzilla.redhat.com/show_bug.cgi?id=1536121
+[rhbz#1542288]: https://bugzilla.redhat.com/show_bug.cgi?id=1542288
 [rhbz#1549535]: https://bugzilla.redhat.com/show_bug.cgi?id=1549535
 [rhbz#1550243]: https://bugzilla.redhat.com/show_bug.cgi?id=1550243
 [rhbz#1557366]: https://bugzilla.redhat.com/show_bug.cgi?id=1557366